requirements.txt
absl_py==2.0.0
chex==0.1.84
distrax==0.1.4
dm_pix==0.4.1
einops==0.7.0
flashbax==0.0.1
flax==0.7.5
gymnasium==1.0.0rc1
hydra-core==1.3.2
ipython==8.18.0
jax==0.4.20
numpy==1.26.2
omegaconf==2.3.0
optax==0.1.7
orbax==0.1.9
pydantic==2.5.2
rich==13.7.0
Shimmy==1.3.0
tqdm==4.66.1
